Add and enable UCA repo by default
When the UCA repo is enabled for nova, a newer version of libvirt will
be installed. The repo_build role will also need to have the UCA repo
enabled to install the same version of libvirt-dev so that a
corresponding libvirt-python wheel will be built.

The test preparation implements an override of the URL for OpenStack-CI
to make use of the local mirrors.

Change-Id: I6d5532e862e98929cff868f940c202875d1ee560