Allow the use of a pre-staged git cache
This patch provides the ability for the repo server to take advantage
of a pre-staged git cache on the deployment host, such as that provided
on OpenStack-CI's nodes.

This reduces the execution time of the build process by:

- only executing a repo clone for invalid or missing repositories
- only updating the clone if the existing clone doesn't already have
  the version required

Due to the way OpenStack-CI stages the cache (from a file based origin
which is not accessible) it is also necessary to adjust the origin
remote instead of creating a new remote as was done previously,
otherwise the 'git fetch' execution would fail.

The patch includes changes to the tests to pre-stage two repositories
and mangle them intentionally to test the resilience of the clone
process.

Change-Id: Ibe70ba27b9a716e9385a5db46cd745bc0d538d63