Correct global_pins depth
In https://review.openstack.org/350352 the repo_build process
implemented the test vars and changes to the constraints build
which caters properly for global pins, but the data set used
was incorrectly mapped from the py_pkgs lookup output.

This patch fixes the depth, tests and template to use the
correct depth.

Related-Bug: #1605846
Closes-Bug: #1609056
Change-Id: Ic8959fd27b4012d569aef1b9ddd1d7be40ed49e0