Only build wheels for services which are deployed
The current repo build process builds wheels for all OpenStack services,
regardless of whether they're actually going to be used.

To reduce the repo build time, this patch will only build wheels for
OpenStack services where there is one or more members in the service
inventory group.

An updated set of test vars is included based on a recent stable/mitaka
output of the py_pkgs lookup based on the below-mentioned dependant
patch.

Depends-On: I5e981c818a1da50aa0ae15630655a8dfac4f3db7
Change-Id: Idda16b4f382eee57c7469af898859d6d81d4eb30