Implement x86 archive links using old name
Until https://review.openstack.org/345986 merges, integrated builds
cannot proceed due to the fact that the repo build is producing
venv archives which are named with the processor architecture in
the name, while the playbooks are expecting a different name.

This patch implements links to the archives using the old names as
a temporary compatibility measure.

Change-Id: Ie0f478aaad1fdddc267c5ad2d4a893701dad78b8