Exclude git sourced packages from upper-constraints test
This patch updates the py_pkgs output based on the current Newton
package set, and to include the new output based on the changes in
https://review.openstack.org/352695

The previous data was based on a lookup with stable/mitaka roles.

As the tempest package is now in upper constraints, the updated data
exposed a flaw in the test which did not exclude git sourced packages
from the upper constraints validation. This patch fixes that.

The results of the tests for this review validate that the old data
structures are intact. A follow on patch will adjust the repo build
process to take advantage of the new data structure.

Change-Id: I231d0f02f3fc370d57518593909c3ed426ea8b4c