Include optional_pip_packages in selective wheel build
The updated py_pkgs lookup in https://review.openstack.org/352695
produces a granular list of all python package requirements in
the roles.

This patch updates the wheel build process to make use of the new
data structure to ensure that the resulting build includes the
optional python packages specified in each role.

It also makes use of the role:host_group data mapping in the new
structure.

Depends-On: Iad00e4f1268187a29e90d7b8e6afb5abfd1e92a2
Change-Id: I695d3f49263f759ece85d23949c9bc560ba1d756