Update repo source download to use processed requirements
Previously the requirements were processed multiple times - once
to build requirements.txt and another to download the pip sources.

This patch ensures that the processed requirements.txt is used to
download the pip sources to ensure that the processing only happens
once and that the downloaded sources are accurate and complete.

Change-Id: I8d056668208a09ec61d318bd67f29cbf32f3b195