Add git cleanup to op-clone-script
This change ensures that the should the `op-clone-script.sh` find a valid
git directory it's working with a clean git history. This is done by simply
running `git clean -f -d` on the target destination. This is important because
there is a chance that a git repository may have had a force push on a given
remote or there were local changes made to the repo that break a new checkout.

Change-Id: I177871a90e08f0d29f28a0923fd8fc89fe91c15c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>