Ensure git folder rights
This patch adds a task to ensure that the git folder on the repo
server has the appropriate use:group rights. This ensures that if
a pre-staged git folder is used, the files and folders all have
the correct rights.

Also, the clone script is no longer removed as this is unnecessary
and is helpful for diagnosis of issues.

Finally, some test changes are made in order to ensure that the
ownership change works properly.

Combined backport of:
- https://review.openstack.org/380553
- https://review.openstack.org/383431

Change-Id: I40a02c26d0abc9d17dd7fad2eab969d169c436bb