Remove warning when no base venvs are found
During the "Destroy base venvs to rebuild them" step ansible tries to
access base_venvs.stdout_lines. If no base_venvs have been registered
previously, this causes a deprecation warning because stdout_lines is
undefined.
This is a small fix to remove this warning by giving an empty list as
default value when stdout_lines is not defined.

TrivialFix

Change-Id: I358709f3d4dc10a60f59976f1c4fbf87705def57