Ensure git folder rights and always run clone script
In https://review.openstack.org/380553 a change was made to
only execute the clone script if the script changes. This
has resulted in a failure when using a pre-staged git cache.

As the script only executes clone tasks if the git repo is
broken or the SHA isn't in the clone, it's OK to execute it
multiple times.

This patch ensures that the script always runs, but also tries
to add changes to the tests to try to uncover the cause of the
integrated build failure which the previous patch caused.

Additionally, the patch also adds tasks to ensure that the git
folder has the correct owner:group.

Change-Id: I40a02c26d0abc9d17dd7fad2eab969d169c436bb