Statically define the repo build concurrency
The repo build concurrency was pulling the processor core count and using
that int as the value for determining concurrent builds. Sadly,
this value can be high, low, or simply undefined and is entirely dependent
on the Ansible fact gathering system which has additional complexities and
dependencies based on how the hardware has represented the information.

Examples:
  * Should the value be low, 3 or less than, the build process would run
    serially making things painfully slow.
  * Should the value be high the build process may run too many tasks at
    one time causing general system instability.

This commit statically defines the concurrency level as "4" which is small
enough to support vagant, and minimal resource machines, while also being
performant enough to support large scale production clouds.

Change-Id: If2ad496c31c4d0092e7856c3deb07035ebc6c8b9