Remove static indexing from repo build process
The repo server has autoindexing enabled, so the tasks
to enumerate the contents of the folder and to produce
a static index file are unnecessary.

This patch removes the tasks and ensures that the files
in an existing repository are also removed to cater for
upgrades.

Change-Id: Ic2bb537f331d14a4ddfefa5ea9ec785d0fce3835
(cherry picked from commit 34b0802a887d12da0cc98f18cc066509860b1250)