Add non *.whl files to MANIFEST.in
This patch switches from using the command/shell module
to using the find module to index the files.

It also adds the files that are not *.whl to the manifest
file to make it easier to implement a staging system for
deployments.

Change-Id: Ie927fb40fdf1f4cb934377320b800a506ed842e2
(cherry picked from commit 5e722bccafd61fa56ffbe7297ae3783c4cf743e9)