Include git clone for packages in requirements
Currently if there is a package in requirements which
needs to be built from one of the git clones, it will
not work unless the related host group is populated.
This is due to the fact that the selective git clone
hinges on the host group population.

This patch adds an extra check for each git clone to
validate whether the package is in the requirements
list too. If it is, then it will clone the git source
regardless of the host group population.

Change-Id: Id121a1bcd91f688c8fab41de0b74b5b8196f8b44
Closes-Bug: #1671958
(cherry picked from commit 7331b821b4e2894a62bd22e7e47e87e04b7d7f71)