Use package module to install distro packages
Consolidate distro package install tasks into a single task using the
package module and pass the package list into the name instead of using
a with_items loop.

The lxc package is no longer a dependency of lxc-dev, so the task to
override lxc-net is no longer necessary and would not effect a systemd
service anyway.

The minimum Ansible version is raised to 2.2 due to a
known bug [1] in Ansible's apt module which does not
update the cache properly if the cache update and the
install are combined in a single task.

[1] https://github.com/ansible/ansible-modules-core/issues/1497

Change-Id: I0b1b5e5413aa3d4496f15f64e3673f4229bbe237