Ensure that all release files are in the manifest
Ansible 2.3 introduced a new option to the find module, but
in doing so they also implemented a breaking change.

Previously the find module would find both files and links
by default, but now it only finds files unless you tell it
specifically to find links.

As all the python wheels in the os-releases folder are links,
we tell the find module to find anything in the folder so
that the manifest file is complete.

Change-Id: Id4ebb2aa6de85a7a2d048bede787240493aadfec
(cherry picked from commit 9ebda8ad089334aae2c497afdc0af92c64aa53f4)