Ensure that all release files are cleaned up when rebuilding
Ansible 2.3 introduced a new option to the find module, but
in doing so they also implemented a breaking change.

Previously the find module would find both files and links
by default, but now it only finds files unless you tell it
specifically to find links.

As all the python wheels in the os-releases folder are links,
we tell the find module to find anything in the folder so
that the cleanup done is complete.

Change-Id: I6a63762f55782d8aeaa2a47aa55da21bc06ab431
(cherry picked from commit d9391e1f68027db2819c8d81e698a01a536c780f)