Correct location of venv build logs
In [1] the intent was to have each venv build log
to its own log file. This makes troubleshooting a
lot easier due to the verbosity of the output in
the log - this at least separates each venv's
output. Unfortunately this extra CLI option was
supposed to be removed. This patch completes the
original intent.

[1] https://review.openstack.org/470757

Change-Id: I714d160a39beb870ba66db11a60cf0241ca770f3
(cherry picked from commit 12611a94408421daecdb7daad18ba857cbaba794)