Remove leftover UCA variables
Previously the repo_build role implemented its own tasks for
setting up the UCA repository. It was later shifted to the
pip_install role and for Queens it's in the openstack_hosts
role.

In Queens these variables are no longer used. In Pike and
before these variables interfere with the values set out in
the pip_install role which are the canonical values.

As such, we remove them from the role and this patch will
need to be ported back.

Change-Id: If499287a5d346bb7c5d0775c21f55a1f73d6db89
(cherry picked from commit 1948d899a5c76143cf24c1a8ae162906a8e1faf1)