Ensure virtualenv is not broken for CentOS/SuSE too
In I2c50bb693dc5dc505bddce34ff45294ac44f7dc9 we applied changes
to the settings to repo_build_venv_command_options so that pip,
setuptools and wheel were not copied into the new virtualend on
creation, but rather installed into it later.

Unfortunately the patch missed that CentOS and SuSE platforms
have an override for the same variable, so the issue set out in
the patch still applies to those platforms.

This patch ensures that the overrides are removed and the
variable is only declared in one place, but the same functionality
which gave rise to the override for those platforms is still
applied when executed.

Closes-Bug: #1746935
Change-Id: Ie0c8c95a0d4315f8d1f848b1309bf88b7c2644d4