Install PBR to ensure we're not breaking builds
A new version of PBR was released which causes some services, like
gnocchi, to break in new and exciting ways. This change installs PBR
within the repo_build role which will ensure all of the correct tools
are available to the repo build process and are installed using known
constraints.

Change-Id: Ibf5e3bf0381ec40b1adbe6ce49a5cf7284d3e147
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>